+++ site/src/components/UserDropdown/UserDropdown.stories.tsx
@@ -32,39 +32,3 @@ ExampleNoRoles.args = {
     return Promise.resolve()
   },
 }
-
-export const ExampleOneRole = Template.bind({})
-ExampleNoRoles.args = {
-  user: {
-    id: "1",
-    username: "CathyCoder",
-    email: "cathy@coder.com",
-    created_at: "dawn",
-    status: "active",
-    organization_ids: [],
-    roles: [{ name: "member", display_name: "Member" }],
-  },
-  onSignOut: () => {
-    return Promise.resolve()
-  },
-}
-
-export const ExampleThreeRoles = Template.bind({})
-ExampleNoRoles.args = {
-  user: {
-    id: "1",
-    username: "CathyCoder",
-    email: "cathy@coder.com",
-    created_at: "dawn",
-    status: "active",
-    organization_ids: [],
-    roles: [
-      { name: "admin", display_name: "Admin" },
-      { name: "member", display_name: "Member" },
-      { name: "auditor", display_name: "Auditor" },
-    ],
-  },
-  onSignOut: () => {
-    return Promise.resolve()
-  },
-}
